backend/pyproject.toml
[project]
name = "kitconcept.intranet"
dynamic = ["version"]
description = "A Plone distribution for Intranets with Plone. Created by kitconcept."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "kitconcept GmbH", email = "info@kitconcept.com" },
]
keywords = ["CMS", "Plone", "Python", "Intranet"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Plone",
    "Framework :: Plone :: 6.1",
    "Framework :: Plone :: Distribution",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "collective.person",
    # "kitconcept.solr",
    "plone.api",
    "plone.distribution>=2.0.0a1",
    "plone.restapi>=9.8.3",
    "plone.volto>=5.0.2",
    "python-dateutil",
]

[project.urls]
Homepage = "https://github.com/kitconcept/kitconcept.intranet"
PyPI = "https://pypi.python.org/pypi/kitconcept.intranet"
Source = "https://github.com/kitconcept/kitconcept.intranet"
Tracker = "https://github.com/kitconcept/kitconcept.intranet/issues"

[project.entry-points."z3c.autoinclude.plugin"]
target = "plone"


[tool.hatch.version]
path = "src/kitconcept/intranet/about.py"

[tool.hatch.build]
strict-naming = false
packages = ["src/kitconcept"]

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.env]
requires = [
    "hatch-pip-compile"
]
[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
python = "3.11"
type = "pip-compile"
path = ".venv"
pip-compile-resolver = "uv"
pip-compile-installer = "uv"
lock-filename = "requirements.lock"
pip-compile-args = [
    "--constraint",
    "constraints-mxdev.txt",
]
dependencies = [
  "coverage[toml]>=6.5",
  "hatch",
  "hatchling==1.24.2",
  "i18ndude",
  "isort",
  "plone.app.testing",
  "plone.restapi[test]",
  "pytest-cov==5.0.0",
  "pytest-docker",
  "pytest-plone>=0.5.0",
  "pytest",
  "ruff",
  "towncrier",
  "zpretty",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]
lint = [
  "ruff check",
  "isort --check-only src/ tests/",
  "zpretty --check src"
]
format = [
  "ruff format",
  "isort src/ tests/",
  "zpretty -i src"
]
i18n = [
  "python -m kitconcept.intranet.locales.update"
]
changelog-draft = "towncrier --draft"

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.coverage.run]
source_pkgs = ["kitconcept.intranet", "tests"]
branch = true
parallel = true
omit = [
  "kitconcept/intranet/locales/__init__.py",
  "kitconcept/intranet/locales/update.py",
]

[tool.coverage.paths]
kitconcept.intranet = ["src/kitconcept/intranet"]
tests = ["tests",]

[tool.coverage.report]
skip_empty = true
show_missing = true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.isort]
profile = "plone"

[tool.towncrier]
package = "kitconcept.intranet"
package_dir = "src"
directory = "news/"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
title_format = "## {version} ({project_date})"
template = "news/.changelog_template.jinja"
underlines = ["", "", ""]
issue_format = "[#{issue}](https://github.com/kitconcept/kitconcept.intranet/issue/{issue})"

[[tool.towncrier.type]]
directory = "breaking"
name = "Breaking"
showcontent = true

[[tool.towncrier.type]]
directory = "feature"
name = "Feature"
showcontent = true

[[tool.towncrier.type]]
directory = "bugfix"
name = "Bugfix"
showcontent = true

[[tool.towncrier.type]]
directory = "internal"
name = "Internal"
showcontent = true

[[tool.towncrier.type]]
directory = "documentation"
name = "Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "test"
name = "Test"
showcontent = true

[tool.ruff]
target-version = "py310"
line-length = 88
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
]
lint.ignore = [
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.codespell]
ignore-words-list = "discreet,vew"
skip = "*.po,*.min.js"

[tool.check-manifest]
ignore = [
    ".editorconfig",
    ".flake8",
    ".meta.toml",
    ".pre-commit-config.yaml",
    "dependabot.yml",
    "mx.ini",
]
